Fixed inconsistencies in the code

+++ Backend/src/main/java/com/mischievous/fairies/persistence/repository/AccountRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface AccountRepository extends JpaRepository<AccountEntity, Integer> {
+public interface AccountRepository extends JpaRepository<AccountEntity, Long> {
     boolean existsByEmail(String email);
     Optional<AccountEntity> findByEmail(String email);
 }
